+++ CHANGES.md
@@ -208,5 +208,6 @@
 ### Pre-release (work-in-progress)
 - Bernafon/Sonic/Philips: target market accepts manually input `default` rather than only accepts pressing Enter with default
 - ReSound/Beltone/Interton: add an option to manually specify a version
+- Widex: option to change target market (country)
 - Clean screen before menus
 - Layout adjustment